+++ assets/javascripts/kitten/karl/pages/comments/comments.stories.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import CommentsPage from './components/comments'
+
+storiesOf('Pages/Comments', module).add('Comments page', () => {
+  return <CommentsPage />
+})